@@ -5,27 +5,25 @@
 </style>
 <template>
     <div class="curation-classification-history">
-        <!-- <pre>{{curation.classifications}}</pre> -->
         <table class="table table-bordered table-small">
             <tr>
                 <th>Classification</th>
                 <th>Date</th>
             </tr>
-            <tr 
-                v-for="(classification, idx) in orderedClassifications" 
-                :key="classification.pivot.id" 
+            <tr
+                v-for="(classification, idx) in orderedClassifications"
+                :key="classification.pivot.id"
                 :class="{'table-primary highlight': (idx == 0)}"
             >
                 <td>{{classification.name}}</td>
-                <td>{{classification.pivot.classification_date | formatDate('YYYY-MM-DD') }}</td>
+                <td>{{ $formatDate(classification.pivot.classification_date, 'YYYY-MM-DD') }}</td>
             </tr>
         </table>
     </div>
 </template>
 <script>
-    import moment from 'moment'
-    import filters from '../../filters'
-    
+    import dayjs from 'dayjs'
+
     export default {
         props: {
             curation: {
@@ -37,7 +35,7 @@
             orderedClassifications: function () {
                 if (this.curation.classifications) {
                     return this.curation.classifications.concat().sort((a, b) => {
-                        if (moment(a.pivot.classification_date).isSame(b.pivot.classification_date)) {
+                        if (dayjs(a.pivot.classification_date).isSame(b.pivot.classification_date)) {
                             if(a.id == b.id) {
                                 return 0
                             }
@@ -46,7 +44,7 @@
                             }
                             return 1;
                         }
-                        if (moment(a.pivot.classification_date).isBefore(b.pivot.classification_date)) {
+                        if (dayjs(a.pivot.classification_date).isBefore(b.pivot.classification_date)) {
                             return 1;
                         }
                         return -1;
