@@ -2,21 +2,16 @@
 
 <template>
     <div class="mb-2">
-        <button 
-            class="btn btn-primary btn-sm" 
+        <button
+            class="btn btn-primary btn-sm"
             @click="showModal = true"
-            
         >Add Document</button>
 
-        <b-modal 
-            v-model="showModal"
-            title="Upload a Document"
-            @ok="uploadFile"
-            @cancel="clearForm"
-            ref="uploadModal"
-            :ok-disabled="uploading"
-            :cancel-disabled="uploading"
-            :ok-title="uploading ? 'Uploading...' : 'Upload'"
+        <Dialog
+            v-model:visible="showModal"
+            header="Upload a Document"
+            modal
+            :style="{ width: '50vw' }"
         >
             <div class="form-row">
                 <label class="col-sm-2" for="file-field">
@@ -26,13 +21,13 @@
                     <div class="d-flex justify-content-between">
                         <div><input type="file" ref="uploadField" class="form-control-file" id="file-field" @change="prepopulateName()" :disabled="uploading"></div>
                         <div>
-                            <small class="text-secondary material-icons cursor-pointer" v-b-toggle.file-info-collapse>info</small>
+                            <small class="text-secondary cursor-pointer" @click="showFileInfo = !showFileInfo">info</small>
                         </div>
                     </div>
-                    <b-collapse id="file-info-collapse">
+                    <div v-show="showFileInfo">
                         <div><small class="text-muted">Supported types: {{supportedMimes.join(', ')}}</small></div>
                         <div><small class="text-muted">Max size: {{maxUploadSize}}</small></div>
-                    </b-collapse>
+                    </div>
                     <validation-error :messages="errors.file"></validation-error>
                 </div>
             </div>
@@ -48,7 +43,7 @@
                 <div class="col-sm-10">
                     <select name="category_id" id="category_id" class="form-control form-control-sm" v-model="newUpload.upload_category_id" :disabled="uploading">
                         <option value="">None</option>
-                        <option 
+                        <option
                             v-for="category in categories"
                             :key="category.id"
                             :value="category.id"
@@ -64,40 +59,51 @@
                     Notes:
                 </label>
                 <div class="col-sm-10">
-                    <textarea 
-                        name="notes" 
-                        v-model="newUpload.notes" 
-                        id="notes" 
-                        cols="30" 
-                        rows="5" 
-                        class="form-control form-control-sm" 
-                        maxlegnth="65535"
+                    <textarea
+                        name="notes"
+                        v-model="newUpload.notes"
+                        id="notes"
+                        cols="30"
+                        rows="5"
+                        class="form-control form-control-sm"
+                        maxlength="65535"
                         :disabled="uploading"
                     ></textarea>
                     <validation-error :messages="errors.notes"></validation-error>
                 </div>
             </div>
-        </b-modal>
+            <template #footer>
+                <button type="button" class="btn btn-secondary" @click="clearForm(); showModal = false" :disabled="uploading">Cancel</button>
+                <button type="button" class="btn btn-primary" @click="uploadFile" :disabled="uploading">
+                    {{ uploading ? 'Uploading...' : 'Upload' }}
+                </button>
+            </template>
+        </Dialog>
     </div>
 </template>
 
 <script>
     import ValidationError from '../../ValidationError.vue'
-    import {mapGetters} from 'vuex'
+    import Dialog from 'primevue/dialog'
+    import { mapState } from 'pinia'
+    import { useAppStore } from '../../../stores/app'
 
     export default {
         components: {
-            ValidationError
+            ValidationError,
+            Dialog
         },
         props: {
             curation: {
                 required: true,
                 type: Object
             }
-        },  
+        },
+        emits: ['uploaded'],
         data() {
             return {
                 showModal: false,
+                showFileInfo: false,
                 categories: [],
                 newUpload: {},
                 errors: {},
@@ -105,7 +111,7 @@
             }
         },
         computed: {
-            ...mapGetters({
+            ...mapState(useAppStore, {
                 maxUploadSize: 'getMaxUploadSize',
                 supportedMimes: 'getSupportedMimes'
             })
@@ -134,9 +140,8 @@
                     this.newUpload.name = this.$refs.uploadField.files[0].name;
                 }
             },
-            uploadFile(evt) {
+            uploadFile() {
                 this.initErrors();
-                evt.preventDefault();
 
                 let formData = new FormData();
                 formData.append('curation_id', this.curation.id);
@@ -148,7 +153,7 @@
                 this.uploading = true;
 
                 return window.axios.post(
-                    `/api/curations/${this.curation.id}/uploads`, 
+                    `/api/curations/${this.curation.id}/uploads`,
                     formData,
                     {
                         headers: {
@@ -159,7 +164,7 @@
                 .then(response => {
                     this.$emit('uploaded');
                     this.clearForm();
-                    this.$nextTick(() => this.$refs.uploadModal.hide());
+                    this.showModal = false;
                 })
                 .catch(error => {
                     if (error.response.status == 422) {
@@ -173,7 +178,7 @@
                     alert('There was an unkown problem with your file upload.');
                 })
                 .then(() => {this.uploading = false})
-        
+
             },
             launchFileSelector() {
                 this.$refs.uploadField.click();
@@ -184,6 +189,6 @@
             this.initNewUpload();
             this.initErrors();
         }
-    
+
 }
 </script>
