@@ -3,15 +3,15 @@
 <template>
     <div>
         <h4>
-            <document-uploader 
-                :curation="curation" 
-                class="float-right"
+            <document-uploader
+                :curation="curation"
+                class="float-end"
                 v-on:uploaded="$refs.docList.getDocuments()"
                 v-if="user.canEditCuration(curation)"
             ></document-uploader>
             Documents
         </h4>
-        
+
         <documents-list :curation="curation" ref="docList"></documents-list>
     </div>
 </template>
@@ -19,7 +19,8 @@
 <script>
     import DocumentUploader from './DocumentUploader.vue';
     import DocumentsList from './DocumentsList.vue';
-    import { mapGetters } from 'vuex';
+    import { mapState } from 'pinia';
+    import { useAppStore } from '../../../stores/app';
 
     export default {
         components: {
@@ -37,7 +38,7 @@
             }
         },
         computed: {
-            ...mapGetters({user: 'getUser'})
+            ...mapState(useAppStore, {user: 'getUser'})
         },
         methods: {
         }
