@@ -2,7 +2,7 @@
 
 <template>
     <div class="component-container w-50">
-        <classification-history :curation="value"></classification-history>
+        <classification-history :curation="updatedCuration"></classification-history>
         <div class="alert alert-secondary">
             Classifications must be added to a curation via the GCI.
         </div>
@@ -11,7 +11,8 @@
 </template>
 
 <script>
-    import { mapGetters, mapActions } from 'vuex'
+    import { mapState, mapActions } from 'pinia'
+    import { useClassificationsStore } from '../../../stores/classifications'
     import curationFormMixin from '../../../mixins/curation_form_mixin'
     import CurationNotifications from './ExistingCurationNotification.vue'
     import DateField from '../../DateField.vue'
@@ -35,18 +36,18 @@
             }
         },
         computed: {
-            ...mapGetters('classifications', {
+            ...mapState(useClassificationsStore, {
                classifications: 'Items',
             }),
         },
         methods: {
-            ...mapActions('classifications', {
+            ...mapActions(useClassificationsStore, {
                 getAllClassifications: 'getAllItems'
-            }),            
+            }),
         },
         mounted: function () {
             this.getAllClassifications()
         }
-    
+
     }
 </script>
