@@ -2,23 +2,22 @@
 
 <template>
     <div>
-        <b-button 
-            variant="info"
-            size="sm" 
-            class="form-control mb-2"
+        <button
+            class="btn btn-info btn-sm form-control mb-2"
             @click="modalVisible = true"
-        >Add or update classification</b-button>
+        >Add or update classification</button>
 
-        <classification-history :curation="value"></classification-history>
+        <classification-history :curation="modelValue"></classification-history>
 
-        <b-modal 
-            v-model="modalVisible"
+        <Dialog
+            v-model:visible="modalVisible"
             @hide="submitAll"
-            size="lg"
+            :style="{ width: '50rem' }"
+            :modal="true"
         >
-            <div slot="modal-header">
+            <template #header>
                 <h3>Update Classification</h3>
-            </div>    
+            </template>
             <table class="table">
                 <thead>
                     <tr>
@@ -29,7 +28,7 @@
                 <tbody>
                     <tr>
                         <td>
-                            <b-form-select v-model="newClassificationId">
+                            <select v-model="newClassificationId" class="form-control">
                                 <option :value="null">Select...</option>
                                 <option v-for="classification in classificationOptions"
                                     :key="classification.id"
@@ -37,92 +36,90 @@
                                 >
                                     {{classification.name}}
                                 </option>
-                            </b-form-select>
+                            </select>
                             <div class="text-danger" v-if="errors.classification_id">
                                 <div v-for="message in errors.classification_id" :key="message"><small>{{message}}</small></div>
                             </div>
                         </td>
-                        <td class="form-inline">
-                            <datepicker 
+                        <td class="d-flex align-items-center">
+                            <DatePicker
                                 v-model="newClassificationDate"
-                                input-class="form-control mr-2"
-                                format='yyyy-MM-dd'
-                                calendar-class="small-calendar"
+                                dateFormat="yy-mm-dd"
                                 placeholder="Select a date"
-                                :highlighted="highlighted"
-                            ></datepicker>
-                            <b-button 
-                                variant="primary"
+                                inputClass="form-control me-2"
+                            />
+                            <button
+                                class="btn btn-primary"
                                 @click="addClassification"
                             >
                                 <strong>+</strong>
-                            </b-button>
+                            </button>
                         </td>
                     </tr>
                     <tr v-for="classification in curationCopy.classifications" :key="classification.pivot.id">
                         <td>
                             <label :for="'classification-date-'+classification.id"><strong>{{classification.name}}</strong></label>
                         </td>
-                        <td class="form-inline">
-                            <datepicker
-                                :id="'classification-date-'+classification.id"
+                        <td class="d-flex align-items-center">
+                            <DatePicker
+                                :inputId="'classification-date-'+classification.id"
                                 v-model="classification.pivot.classification_date"
-                                input-class="form-control mr-2"
-                                format='yyyy-MM-dd'
-                                calendar-class="small-calendar"
+                                dateFormat="yy-mm-dd"
                                 placeholder="Select a date"
-                                :highlighted="highlighted"
-                                @selected="updateclassificationDate(classification.pivot,$event)"
-                            ></datepicker>
-                            <b-button @click="removeclassificationEntry(classification)"><strong>x</strong></b-button>
+                                inputClass="form-control me-2"
+                                @date-select="updateclassificationDate(classification.pivot, $event)"
+                            />
+                            <button class="btn btn-secondary" @click="removeclassificationEntry(classification)"><strong>x</strong></button>
                         </td>
                     </tr>
                 </tbody>
             </table>
-        </b-modal>
+        </Dialog>
     </div>
 </template>
 
 <script>
-    import { mapGetters, mapActions } from 'vuex'
+    import { mapState, mapActions } from 'pinia'
+    import { useClassificationsStore } from '../../../stores/classifications'
+    import { useCurationsStore } from '../../../stores/curations'
     import ClassificationHistory from '../ClassificationHistory.vue'
-    import Datepicker from 'vuejs-datepicker'
-    import moment from 'moment'
+    import DatePicker from 'primevue/datepicker'
+    import Dialog from 'primevue/dialog'
+    import dayjs from 'dayjs'
+    import { formatDate } from '../../../utils/formatDate'
 
     export default {
         components: {
             ClassificationHistory,
-            Datepicker
+            DatePicker,
+            Dialog,
         },
         props: {
-            value: {
+            modelValue: {
                 required: true,
                 type: Object
             }
         },
+        emits: ['update:modelValue'],
         data() {
             return {
                 curationCopy: {},
                 modalVisible: false,
                 newClassificationDate: new Date(),
                 newClassificationId: null,
-                highlighted: {
-                    from: new moment().hour(0),
-                    to: new moment().hour(24)
-                },
                 classificationDatesUpdated: false,
                 errors: {}
             }
         },
         watch: {
-            value: {
+            modelValue: {
                 handler: 'syncCuration',
                 immediate: true,
                 deep: true
             }
         },
         computed: {
-            ...mapGetters('classifications', {
+            ...mapState(useClassificationsStore, {
                 classifications: 'Items',
             }),
             classificationOptions() {
@@ -130,7 +127,7 @@
             },
         },
         methods: {
-            ...mapActions('curations', {
+            ...mapActions(useCurationsStore, {
                 linkNewClassification: 'linkNewClassification',
                 updateClassification: 'updateClassification',
                 unlinkClassification: 'unlinkClassification'
@@ -138,10 +135,10 @@
             addClassification() {
                 this.linkNewClassification(
                     {
-                        curation: this.curationCopy, 
+                        curation: this.curationCopy,
                         data: {
                             classification_id: this.newClassificationId,
-                            classification_date: this.$options.filters.formatDate(this.newClassificationDate, 'YYYY-MM-DD')
+                            classification_date: formatDate(this.newClassificationDate, 'YYYY-MM-DD')
                         }
                     }
                 ).then(response => {
@@ -153,7 +150,7 @@
                 })
             },
             updateclassificationDate(pivot, newDate) {
-                if (!pivot || moment(pivot.classification_date).diff(newDate) == 0) {
+                if (!pivot || dayjs(pivot.classification_date).diff(newDate) == 0) {
                     return;
                 }
                 this.updateClassification({
@@ -161,10 +158,7 @@
                     pivotId: pivot.id,
                     data: {
                         classification_id: pivot.classification_id,
-                        classification_date: moment(newDate).format('YYYY-MM-DD')}
-                })
-                .then(response => {
-                    console.log('classification date updated')
+                        classification_date: dayjs(newDate).format('YYYY-MM-DD')}
                 })
                 .catch(response => {
                     this.errors = response.data.errors;
@@ -180,7 +174,7 @@
                 }
             },
             syncCuration() {
-                this.curationCopy = JSON.parse(JSON.stringify(this.value))
+                this.curationCopy = JSON.parse(JSON.stringify(this.modelValue))
             },
         },
 
