@@ -1,13 +1,11 @@
 <style scoped>
     .unused {
         color: #aaa;
-        /*text-decoration: line-through;*/
     }
 </style>
 <template>
     <div class="component-container row">
         <div v-if="phenotypes.length > 0" class=" col-lg-8">
-            <!-- <strong>In this curation</strong> -->
             <table class="table table-sm table-xs mb-0">
                 <thead>
                     <th>MIM Number</th>
@@ -27,13 +25,11 @@
     </div>
 </template>
 <script>
-    import OmimRepo from './../../../repositories/OmimRepository';
-
     export default {
         props: {
             geneSymbol: {
                 required: true,
-            }, 
+            },
             curation: {
                 required: true,
                 type: Object
