@@ -1,64 +1,58 @@
 <script>
-import moment from 'moment';
-import {LineToLineMappedSource} from 'webpack-sources';
-import {mapActions} from 'vuex'
+import dayjs from 'dayjs';
+import { mapActions } from 'pinia'
+import { useCurationsStore } from '../../stores/curations'
+import { useAppStore } from '../../stores/app'
 import queryStringFromParams from '../../http/query_string_from_params';
 
 export default {
-    name: 'ComponentName',
+    name: 'SendToGciButton',
     props: {
         curation: {
             type: Object,
-            required: LineToLineMappedSource
+            required: true
         }
     },
+    emits: ['saved'],
     data() {
         return {
-            
+
         }
     },
     computed: {
+        sendToGciEnabled() {
+            const appStore = useAppStore()
+            return appStore.features.sendToGciEnabled
+        },
         enabled: function () {
-            return this.curation.hgnc_id 
-                && this.curation.disease 
+            return this.curation.hgnc_id
+                && this.curation.disease
                 && this.curation.mode_of_inheritance
                 && !this.curation.gdm_uuid;
         },
         popoverText () {
             if (!this.enabled) {
-                const reason = (this.curation.gdm_uuid) 
-                    ? 'the curation is already associatd with a GCI record.' 
+                const reason = (this.curation.gdm_uuid)
+                    ? 'the curation is already associatd with a GCI record.'
                     : ' the curation is not complete.';
                 return `Disabled because ${reason}`
             }
             return null;
         }
     },
-    watch: {
-        curation: {
-            deep: true,
-            immediate: true,
-            handler: function () {
-            }
-        }
-    },
     methods: {
-        ...mapActions('curations', {
-            // fetchCuration: 'fetchItem',
-            // storeNewItem: 'storeNewItem',
+        ...mapActions(useCurationsStore, {
             storeItemUpdates: 'storeItemUpdates',
             linkNewStatus: 'linkNewStatus',
         }),
         async handleClick () {
             await this.storeItemUpdates(this.curation);
             await this.linkNewStatus({
-                    curation: this.curation, 
+                    curation: this.curation,
                     data: {
                         curation_status_id: 4,
-                        status_date: moment().format('YYYY-MM-DD')
+                        status_date: dayjs().format('YYYY-MM-DD')
                     }
-                }).then(rsp => {
-                    console.log('should have linked new status');
                 });
             this.$emit('saved');
 
@@ -75,7 +69,6 @@ export default {
             }
 
             const url = `https://curation.clinicalgenome.org/create-gene-disease${queryStringFromParams(params)}`
-            console.log(url);
             window.open(url, '_gci');
         }
     }
@@ -83,18 +76,15 @@ export default {
 </script>
 
 <template>
-    <div v-if="$store.state.features.sendToGciEnabled">
-        <span id="send-to-gci-button">            
-            <button class="btn btn-primary btn-lg" 
-                :disabled="!enabled" 
+    <div v-if="sendToGciEnabled">
+        <span id="send-to-gci-button">
+            <button class="btn btn-primary btn-lg"
+                :disabled="!enabled"
                 @click="handleClick"
                 :title="popoverText"
             >
                 Complete PreCuration and Go to GCI
             </button>
         </span>
-        <b-popover target="send-to-gci-button" triggers="hover" placement="top" v-if="!enabled">
-            {{popoverText}}
-        </b-popover>
     </div>
 </template>
