@@ -6,107 +6,130 @@
                 <label :for="searchFieldId">Search:</label>&nbsp;
                 <select name="" id="" v-model="filterField" class="form-control form-control-sm">
                     <option :value="null">Any Field</option>
-                    <option :value="field.key" v-for="field in filterableFields" :key="field.name">{{field.label}}</option>
+                    <option :value="field.key" v-for="field in filterableFields" :key="field.key">{{field.label}}</option>
                 </select>
                 &nbsp;
                 <input v-model="filter" placeholder="search curations" class="form-control form-control-sm" :id="searchFieldId" />
             </div>
             <div class="col-md-6">
-                <b-pagination 
-                    size="sm" 
-                    hide-goto-end-buttons 
-                    :total-rows="totalRows" 
-                    :per-page="pageLength" 
-                    v-model="currentPage" 
-                    class="curations-table-pagination my-0 float-right" />    
+                <Paginator
+                    :rows="pageLength"
+                    :totalRecords="totalRows"
+                    :first="(currentPage - 1) * pageLength"
+                    @page="onPage"
+                    class="float-end my-0"
+                    template="PrevPageLink CurrentPageReport NextPageLink"
+                />
             </div>
         </div>
         <div v-show="loading" class="text-center">
             <p class="lead">loading...</p>
         </div>
-        <b-table striped hover 
+        <DataTable
+            :value="items"
+            :loading="tableLoading"
+            lazy
+            :totalRecords="totalRows"
+            :rows="pageLength"
+            :sortField="sortKey"
+            :sortOrder="sortOrder"
+            @sort="onSort"
+            stripedRows
+            showGridlines
             ref="table"
-            :items="curationProvider" 
-            :fields="fields" 
-            :filter="filter"
-            :per-page="pageLength"
-            :current-page="currentPage"
-            @sort-changed="handleSortChanged"
-            :sort-by.sync="sortKey"
-            :sort-desc.sync="sortDesc"
-            :no-local-sorting="true"
-            :show-empty="true"
-        >     
-            <template v-slot:table-busy>
-                <center>Loading...</center>
-            </template>
-            <template v-slot:cell(gene_symbol)="{item}">
-                <router-link
-                    :id="'show-curation-'+item.id+'-link'" 
-                    :to="'/curations/'+item.id"
-                >
-                    {{item.gene_symbol}}
-                </router-link>
-                <small v-if="item.hgnc_id">(hgnc:{{item.hgnc_id}})</small>
-            </template>
-            <template v-slot:cell(mode_of_inheritance)="{item}">
-                <div v-if="item.mode_of_inheritance !== null">
-                    <div :title="item.mode_of_inheritance.name">
-                        {{item.mode_of_inheritance.abbreviation}}
-                    </div>
-                </div>
-            </template>
-            <template v-slot:cell(expert_panel)="{item}">
-                <div>{{(item.expert_panel) ? item.expert_panel.name : null}}</div>
-            </template>
-            <template v-slot:cell(curator)="{item}">
-                <div>{{(item.curator) ? item.curator.name : null}}</div>
-            </template>
-            <template v-slot:cell(current_status)="{item}">
-                <div>{{(item.current_status) ? item.current_status.name : null}}</div>
-            </template>
-            <template v-slot:cell(mondo_id)="{item}">
-                <div>{{ getDiseaseEntityColumn(item) }}</div>
-            </template>
-            <template v-slot:cell(actions)="{item}" class="text-right">
-                <div>
+        >
+            <template #empty>No curations found.</template>
+            <template #loading>Loading...</template>
+
+            <Column field="gene_symbol" header="Gene Symbol" sortable>
+                <template #body="{data}">
                     <router-link
-                        v-if="user.canEditCuration(item)"
-                        :id="'edit-curation-'+item.id+'-btn'" 
-                        class="btn btn-secondary btn-sm" 
-                        :to="'/curations/'+item.id+'/edit'"
+                        :id="'show-curation-'+data.id+'-link'"
+                        :to="'/curations/'+data.id"
                     >
-                        Edit
+                        {{data.gene_symbol}}
                     </router-link>
-                    <delete-button :curation="item" class="btn-sm">
-                        <span class="fa fa-trash">X</span>
-                    </delete-button>
-                </div>
-            </template>
-        </b-table>
+                    <small v-if="data.hgnc_id">(hgnc:{{data.hgnc_id}})</small>
+                </template>
+            </Column>
+            <Column field="mode_of_inheritance" header="MOI" sortable>
+                <template #body="{data}">
+                    <div v-if="data.mode_of_inheritance !== null">
+                        <div :title="data.mode_of_inheritance.name">
+                            {{data.mode_of_inheritance.abbreviation}}
+                        </div>
+                    </div>
+                </template>
+            </Column>
+            <Column field="mondo_id" header="Disease Entity" sortable style="width: 9rem">
+                <template #body="{data}">
+                    <div>{{ getDiseaseEntityColumn(data) }}</div>
+                </template>
+            </Column>
+            <Column field="expert_panel" header="Expert Panel" sortable>
+                <template #body="{data}">
+                    <div>{{(data.expert_panel) ? data.expert_panel.name : null}}</div>
+                </template>
+            </Column>
+            <Column field="curator" header="Curator" sortable>
+                <template #body="{data}">
+                    <div>{{(data.curator) ? data.curator.name : null}}</div>
+                </template>
+            </Column>
+            <Column field="current_status" header="Status" style="width: 8rem">
+                <template #body="{data}">
+                    <div>{{(data.current_status) ? data.current_status.name : null}}</div>
+                </template>
+            </Column>
+            <Column field="id" header="Precuration ID" sortable></Column>
+            <Column header="" style="width: 7rem">
+                <template #body="{data}">
+                    <div>
+                        <router-link
+                            v-if="user.canEditCuration(data)"
+                            :id="'edit-curation-'+data.id+'-btn'"
+                            class="btn btn-secondary btn-sm"
+                            :to="'/curations/'+data.id+'/edit'"
+                        >
+                            Edit
+                        </router-link>
+                        <delete-button :curation="data" class="btn-sm">
+                            <span class="fa fa-trash">X</span>
+                        </delete-button>
+                    </div>
+                </template>
+            </Column>
+        </DataTable>
         <div class="row border-top pt-4">
             <div class="col-md-6">Total Records: {{totalRows}}</div>
             <div class="col-md-6">
-                <b-pagination 
-                    size="sm" 
-                    hide-goto-end-buttons 
-                    :total-rows="totalRows" 
-                    :per-page="pageLength" 
-                    v-model="currentPage" 
-                    class="curations-table-pagination my-0 float-right" />    
+                <Paginator
+                    :rows="pageLength"
+                    :totalRecords="totalRows"
+                    :first="(currentPage - 1) * pageLength"
+                    @page="onPage"
+                    class="float-end my-0"
+                    template="PrevPageLink CurrentPageReport NextPageLink"
+                />
             </div>
         </div>
     </div>
 </template>
 <script>
+    import DataTable from 'primevue/datatable'
+    import Column from 'primevue/column'
+    import Paginator from 'primevue/paginator'
     import getPageOfCurations from '../../resources/curations/get_page_of_curations'
     import uniqid from '../../helpers/uniqid'
-    import { mapGetters } from 'vuex'
+    import { mapState } from 'pinia'
+    import { useAppStore } from '../../stores/app'
     import DeleteButton from './DeleteButton.vue'
-
 
     export default {
         components: {
+            DataTable,
+            Column,
+            Paginator,
             DeleteButton
         },
         props: {
@@ -138,6 +161,8 @@
                 sortKey: JSON.parse(JSON.stringify(this.sortBy)),
                 totalRows: 0,
                 searchFieldId: `search-filter-${uniqid()}`,
+                items: [],
+                tableLoading: false,
                 fields: [
                     {
                         key: 'gene_symbol',
@@ -156,9 +181,6 @@
                         label: 'Disease Entity',
                         sortable: true,
                         filterable: true,
-                        thStyle: {
-                            width: "9rem"
-                        }
                     },
                     {
                         key: 'expert_panel',
@@ -174,12 +196,7 @@
                     {
                         key: 'current_status',
                         label: 'Status',
-                        // sortable: true,
                         sortable: false,
-                        // filterable: true,
-                        thStyle: {
-                            width: "8rem"
-                        }
                     },
                     {
                         key: 'id',
@@ -190,48 +207,64 @@
                         key: 'actions',
                         label: '',
                         sortable: false,
-                        thStyle: {
-                            width: "7rem"
-                        }
                     }
                 ],
-                ctx: null
             }
         },
         computed: {
-            ...mapGetters({user: 'getUser'}),
+            ...mapState(useAppStore, {user: 'getUser'}),
             loading: function () {
                 return false;
             },
             filterableFields() {
                 return this.fields.filter(f => f.filterable)
+            },
+            sortOrder() {
+                return this.sortDesc ? -1 : 1;
             }
         },
         watch: {
             filter: function (to, from) {
                 if (to != from) {
                     this.resetCurrentPage();
+                    this.loadData();
                 }
             },
             filterField: function () {
-                this.$refs.table.refresh()
+                this.loadData();
             }
         },
         methods: {
-            curationProvider(ctx, callback) {
-                if (ctx == this.ctx) {
-                    return;
-                }
-
-                const context = {...ctx, ...this.searchParams, ...{filter_field: this.filterField}};
-                if (this.filterField) {
-                    context.filter_field = this.filterField;
-                }
+            loadData() {
+                this.tableLoading = true;
+                const context = {
+                    currentPage: this.currentPage,
+                    perPage: this.pageLength,
+                    sortBy: this.sortKey,
+                    sortDesc: this.sortDesc,
+                    filter: this.filter,
+                    filter_field: this.filterField,
+                    ...this.searchParams,
+                };
                 getPageOfCurations(context)
                     .then(response => {
-                        this.totalRows = response.data.meta.total
-                        callback(response.data.data)
+                        this.totalRows = response.data.meta.total;
+                        this.items = response.data.data;
+                        this.tableLoading = false;
                     })
+                    .catch(() => {
+                        this.tableLoading = false;
+                    });
+            },
+            onPage(event) {
+                this.currentPage = event.page + 1;
+                this.loadData();
+            },
+            onSort(event) {
+                this.sortKey = event.sortField;
+                this.sortDesc = event.sortOrder === -1;
+                this.resetCurrentPage();
+                this.loadData();
             },
             resetCurrentPage () {
                 this.currentPage = 1;
@@ -251,13 +284,12 @@
 
                 return null
             },
-            handleFiltered () {
-              // Trigger pagination to update the number of buttons/pages due to filtering
-              this.resetCurrentPage();
-            },
             handleSortChanged() {
                 this.resetCurrentPage();
             }
+        },
+        mounted() {
+            this.loadData();
         }
     }
 </script>
