@@ -2,7 +2,6 @@
     <div>
         <h4 class="d-flex justify-content-between">
             <div><slot name="title">Notes</slot></div>
-            <!-- <button class="btn btn-sm btn-primary" @click="showAddForm = true">Add Note</button> -->
         </h4>
         <div v-if="notes.length == 0" class="alert alert-light border">
             Not administrative notes for this record.
@@ -19,22 +18,15 @@
                 <tr v-for="note in notes" :key="note.id">
                     <td>{{note.topic ? note.topic.toUpperCase()+': ' : ''}} {{note.content}}</td>
                     <td>{{note.author.name}}</td>
-                    <td>{{note.created_at | formatDate}}</td>
+                    <td>{{$formatDate(note.created_at)}}</td>
                 </tr>
             </tbody>
         </table>
-        <!-- <b-modal v-model="showAddForm" title="Add a note">
-            <textarea style="width: 100%; height: 10rem;" class="form-control" v-model="newNoteContent"></textarea>
-        </b-modal> -->
     </div>
 </template>
 <script>
-import InputRow from './forms/InputRow.vue'
 export default {
     name: 'NotesList',
-    components: {
-        InputRow
-    },
     props: {
         notes: {
             type: Array,
