@@ -2,19 +2,21 @@
 
 <template>
     <transition name="fade">
-        <div class="alert alert-info" v-show="$store.getters.omimLoading">
+        <div class="alert alert-info" v-show="omimLoading">
             Loading data from OMIM...
         </div>
     </transition>
 </template>
 
 <script>
+    import { useAppStore } from '../stores/app'
+
     export default {
-        components: {},
-        data() {
-            return {}
-        },
-        methods: {}
-    
-}
+        computed: {
+            omimLoading() {
+                const appStore = useAppStore()
+                return appStore.omimLoading
+            }
+        }
+    }
 </script>
