@@ -1,5 +1,6 @@
 <script>
-    import {mapGetters, mapActions} from 'vuex'
+    import { mapState, mapActions } from 'pinia'
+    import { useWorkingGroupsStore } from '../../stores/workingGroups'
     import ExpertPanelTabs from './ExpertPanelTabs.vue'
 
     export default {
@@ -14,11 +15,12 @@
         data() {
             return {
                 hasPanels: false,
-                loading: false
+                loading: false,
+                activePanel: 0,
             };
         },
         computed: {
-            ...mapGetters('workingGroups', {
+            ...mapState(useWorkingGroupsStore, {
                 groups: 'Items',
                 getGroup: 'getItemById'
             }),
@@ -27,13 +29,13 @@
                     return {};
                 }
                 const group = this.getGroup(this.id)
-                
+
                 this.hasPanels = group && group.expert_panels && group.expert_panels.length > 0
                 return group
             },
         },
         methods: {
-            ...mapActions('workingGroups', {
+            ...mapActions(useWorkingGroupsStore, {
                 fetchGroup: 'fetchItem'
             }),
         },
@@ -54,24 +56,41 @@
     <div>
         <div>
             <router-link to="/working-groups">
-                Working groups 
+                Working groups
             </router-link>
             &gt;
             <router-link :to="`/working-groups/${group.id}`">
                 {{group.name}}
             </router-link>
         </div>
-        <div class="Working group detail card">
+        <div class="card">
             <div class="card-header">
                 <h3>{{group.name}}</h3>
             </div>
             <div class="card-body">
                 <h4>Expert Panels</h4>
-                <b-tabs pills card vertical v-show="hasPanels" nav-wrapper-class="w-25">
-                    <b-tab v-for="panel in group.expert_panels" :key="panel.id" :title="panel.name">
-                        <ExpertPanelTabs :expert-panel="panel" />
-                    </b-tab>
-                </b-tabs>
+                <div class="d-flex" v-show="hasPanels">
+                    <div class="nav flex-column nav-pills me-3" style="min-width: 25%;" role="tablist" aria-orientation="vertical">
+                        <button
+                            v-for="(panel, index) in group.expert_panels"
+                            :key="panel.id"
+                            class="nav-link"
+                            :class="{ active: activePanel === index }"
+                            @click="activePanel = index"
+                        >
+                            {{ panel.name }}
+                        </button>
+                    </div>
+                    <div class="flex-grow-1">
+                        <div
+                            v-for="(panel, index) in group.expert_panels"
+                            :key="panel.id"
+                            v-show="activePanel === index"
+                        >
+                            <ExpertPanelTabs :expert-panel="panel" />
+                        </div>
+                    </div>
+                </div>
                 <div class="alert alert-secondary" v-show="!hasPanels && !loading">
                     This working group does not have any expert panels
                 </div>
