@@ -10,7 +10,7 @@
 <script>
 export default {
     props: {
-        value: {
+        modelValue: {
             required: true,
             type: Boolean
         },
@@ -23,19 +23,15 @@ export default {
             default: 'Hide'
         }
     },
-    data() {
-        return {
-            
-        }
-    },
+    emits: ['update:modelValue'],
     computed: {
         otherStateLabel () {
-            return this.value ? this.hideLabel : this.showLabel;
+            return this.modelValue ? this.hideLabel : this.showLabel;
         }
     },
     methods: {
         toggle () {
-            this.$emit('input', !this.value);
+            this.$emit('update:modelValue', !this.modelValue);
         }
     }
 }
