@@ -1,8 +1,8 @@
 <template>
     <div>
-        <input 
-            type="date" 
-            :value="formattedDate" 
+        <input
+            type="date"
+            :value="formattedDate"
             class="form-control"
             @input="setDate"
         >
@@ -11,35 +11,25 @@
 <script>
 export default {
     props: {
-        value: {
+        modelValue: {
             required: false,
             default: null
         }
     },
-    emits: [
-        'update:modelValue'
-    ],
-    data() {
-        return {
-            
-        }
-    },
+    emits: ['update:modelValue'],
     computed: {
         formattedDate () {
-            if (!this.value) {
+            if (!this.modelValue) {
                 return null;
             }
-            const fmtdt = this.formatDate(this.value)
-            console.log(fmtdt);
-            return fmtdt
+            return this.formatDate(this.modelValue)
         }
     },
     methods: {
         setDate(event) {
             const date = new Date(Date.parse(event.target.value));
             const adjustedDate = new Date(date.getTime() + date.getTimezoneOffset()*60*1000);
-
-            this.$emit('input', adjustedDate)
+            this.$emit('update:modelValue', adjustedDate)
         },
         formatDate(date) {
             var d = new Date(date),
@@ -47,14 +37,13 @@ export default {
                 day = '' + d.getDate(),
                 year = d.getFullYear();
 
-            if (month.length < 2) 
+            if (month.length < 2)
                 month = '0' + month;
-            if (day.length < 2) 
+            if (day.length < 2)
                 day = '0' + day;
 
             return [year, month, day].join('-');
         }
-
     }
 }
 </script>
