@@ -22,7 +22,7 @@
                         </thead>
 
                         <tbody>
-                            <tr v-for="token in tokens">
+                            <tr v-for="token in tokens" :key="token.id">
                                 <!-- Client Name -->
                                 <td style="vertical-align: middle;">
                                     {{ token.client.name }}
@@ -52,40 +52,17 @@
 
 <script>
     export default {
-        /*
-         * The component's data.
-         */
         data() {
             return {
                 tokens: []
             };
         },
 
-        /**
-         * Prepare the component (Vue 1.x).
-         */
-        ready() {
-            this.prepareComponent();
-        },
-
-        /**
-         * Prepare the component (Vue 2.x).
-         */
         mounted() {
-            this.prepareComponent();
+            this.getTokens();
         },
 
         methods: {
-            /**
-             * Prepare the component (Vue 2.x).
-             */
-            prepareComponent() {
-                this.getTokens();
-            },
-
-            /**
-             * Get all of the authorized tokens for the user.
-             */
             getTokens() {
                 axios.get('/oauth/tokens')
                         .then(response => {
@@ -93,9 +70,6 @@
                         });
             },
 
-            /**
-             * Revoke the given token.
-             */
             revoke(token) {
                 axios.delete('/oauth/tokens/' + token.id)
                         .then(response => {
