@@ -37,7 +37,7 @@
                     </thead>
 
                     <tbody>
-                        <tr v-for="client in clients">
+                        <tr v-for="client in clients" :key="client.id">
                             <!-- ID -->
                             <td style="vertical-align: middle;">
                                 {{ client.id }}
@@ -73,151 +73,127 @@
         </div>
 
         <!-- Create Client Modal -->
-        <div class="modal fade" id="modal-create-client" tabindex="-1" role="dialog">
-            <div class="modal-dialog">
-                <div class="modal-content">
-                    <div class="modal-header">
-                        <h4 class="modal-title">
-                            Create Client
-                        </h4>
+        <Dialog v-model:visible="showCreateModal" header="Create Client" :modal="true" :style="{ width: '500px' }">
+            <!-- Form Errors -->
+            <div class="alert alert-danger" v-if="createForm.errors.length > 0">
+                <p class="mb-0"><strong>Whoops!</strong> Something went wrong!</p>
+                <br>
+                <ul>
+                    <li v-for="(error, idx) in createForm.errors" :key="idx">
+                        {{ error }}
+                    </li>
+                </ul>
+            </div>
 
-                        <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
-                    </div>
+            <!-- Create Client Form -->
+            <form role="form">
+                <!-- Name -->
+                <div class="mb-3 row">
+                    <label class="col-md-3 col-form-label">Name</label>
 
-                    <div class="modal-body">
-                        <!-- Form Errors -->
-                        <div class="alert alert-danger" v-if="createForm.errors.length > 0">
-                            <p class="mb-0"><strong>Whoops!</strong> Something went wrong!</p>
-                            <br>
-                            <ul>
-                                <li v-for="error in createForm.errors">
-                                    {{ error }}
-                                </li>
-                            </ul>
-                        </div>
+                    <div class="col-md-9">
+                        <input id="create-client-name" type="text" class="form-control"
+                                                    @keyup.enter="store" v-model="createForm.name" ref="createNameInput">
 
-                        <!-- Create Client Form -->
-                        <form role="form">
-                            <!-- Name -->
-                            <div class="form-group row">
-                                <label class="col-md-3 col-form-label">Name</label>
-
-                                <div class="col-md-9">
-                                    <input id="create-client-name" type="text" class="form-control"
-                                                                @keyup.enter="store" v-model="createForm.name">
-
-                                    <span class="form-text text-muted">
-                                        Something your users will recognize and trust.
-                                    </span>
-                                </div>
-                            </div>
-
-                            <!-- Redirect URL -->
-                            <div class="form-group row">
-                                <label class="col-md-3 col-form-label">Redirect URL</label>
-
-                                <div class="col-md-9">
-                                    <input type="text" class="form-control" name="redirect"
-                                                    @keyup.enter="store" v-model="createForm.redirect">
-
-                                    <span class="form-text text-muted">
-                                        Your application's authorization callback URL.
-                                    </span>
-                                </div>
-                            </div>
-                        </form>
-                    </div>
-
-                    <!-- Modal Actions -->
-                    <div class="modal-footer">
-                        <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-
-                        <button type="button" class="btn btn-primary" @click="store">
-                            Create
-                        </button>
+                        <span class="form-text text-muted">
+                            Something your users will recognize and trust.
+                        </span>
                     </div>
                 </div>
-            </div>
-        </div>
+
+                <!-- Redirect URL -->
+                <div class="mb-3 row">
+                    <label class="col-md-3 col-form-label">Redirect URL</label>
+
+                    <div class="col-md-9">
+                        <input type="text" class="form-control" name="redirect"
+                                        @keyup.enter="store" v-model="createForm.redirect">
+
+                        <span class="form-text text-muted">
+                            Your application's authorization callback URL.
+                        </span>
+                    </div>
+                </div>
+            </form>
+
+            <!-- Modal Actions -->
+            <template #footer>
+                <button type="button" class="btn btn-secondary" @click="showCreateModal = false">Close</button>
+
+                <button type="button" class="btn btn-primary" @click="store">
+                    Create
+                </button>
+            </template>
+        </Dialog>
 
         <!-- Edit Client Modal -->
-        <div class="modal fade" id="modal-edit-client" tabindex="-1" role="dialog">
-            <div class="modal-dialog">
-                <div class="modal-content">
-                    <div class="modal-header">
-                        <h4 class="modal-title">
-                            Edit Client
-                        </h4>
+        <Dialog v-model:visible="showEditModal" header="Edit Client" :modal="true" :style="{ width: '500px' }">
+            <!-- Form Errors -->
+            <div class="alert alert-danger" v-if="editForm.errors.length > 0">
+                <p class="mb-0"><strong>Whoops!</strong> Something went wrong!</p>
+                <br>
+                <ul>
+                    <li v-for="(error, idx) in editForm.errors" :key="idx">
+                        {{ error }}
+                    </li>
+                </ul>
+            </div>
 
-                        <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
-                    </div>
+            <!-- Edit Client Form -->
+            <form role="form">
+                <!-- Name -->
+                <div class="mb-3 row">
+                    <label class="col-md-3 col-form-label">Name</label>
 
-                    <div class="modal-body">
-                        <!-- Form Errors -->
-                        <div class="alert alert-danger" v-if="editForm.errors.length > 0">
-                            <p class="mb-0"><strong>Whoops!</strong> Something went wrong!</p>
-                            <br>
-                            <ul>
-                                <li v-for="error in editForm.errors">
-                                    {{ error }}
-                                </li>
-                            </ul>
-                        </div>
+                    <div class="col-md-9">
+                        <input id="edit-client-name" type="text" class="form-control"
+                                                    @keyup.enter="update" v-model="editForm.name" ref="editNameInput">
 
-                        <!-- Edit Client Form -->
-                        <form role="form">
-                            <!-- Name -->
-                            <div class="form-group row">
-                                <label class="col-md-3 col-form-label">Name</label>
-
-                                <div class="col-md-9">
-                                    <input id="edit-client-name" type="text" class="form-control"
-                                                                @keyup.enter="update" v-model="editForm.name">
-
-                                    <span class="form-text text-muted">
-                                        Something your users will recognize and trust.
-                                    </span>
-                                </div>
-                            </div>
-
-                            <!-- Redirect URL -->
-                            <div class="form-group row">
-                                <label class="col-md-3 col-form-label">Redirect URL</label>
-
-                                <div class="col-md-9">
-                                    <input type="text" class="form-control" name="redirect"
-                                                    @keyup.enter="update" v-model="editForm.redirect">
-
-                                    <span class="form-text text-muted">
-                                        Your application's authorization callback URL.
-                                    </span>
-                                </div>
-                            </div>
-                        </form>
-                    </div>
-
-                    <!-- Modal Actions -->
-                    <div class="modal-footer">
-                        <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-
-                        <button type="button" class="btn btn-primary" @click="update">
-                            Save Changes
-                        </button>
+                        <span class="form-text text-muted">
+                            Something your users will recognize and trust.
+                        </span>
                     </div>
                 </div>
-            </div>
-        </div>
+
+                <!-- Redirect URL -->
+                <div class="mb-3 row">
+                    <label class="col-md-3 col-form-label">Redirect URL</label>
+
+                    <div class="col-md-9">
+                        <input type="text" class="form-control" name="redirect"
+                                        @keyup.enter="update" v-model="editForm.redirect">
+
+                        <span class="form-text text-muted">
+                            Your application's authorization callback URL.
+                        </span>
+                    </div>
+                </div>
+            </form>
+
+            <!-- Modal Actions -->
+            <template #footer>
+                <button type="button" class="btn btn-secondary" @click="showEditModal = false">Close</button>
+
+                <button type="button" class="btn btn-primary" @click="update">
+                    Save Changes
+                </button>
+            </template>
+        </Dialog>
     </div>
 </template>
 
 <script>
+    import Dialog from 'primevue/dialog'
+
     export default {
-        /*
-         * The component's data.
-         */
+        components: {
+            Dialog,
+        },
         data() {
             return {
                 clients: [],
+                showCreateModal: false,
+                showEditModal: false,
 
                 createForm: {
                     errors: [],
@@ -233,39 +209,11 @@
             };
         },
 
-        /**
-         * Prepare the component (Vue 1.x).
-         */
-        ready() {
-            this.prepareComponent();
-        },
-
-        /**
-         * Prepare the component (Vue 2.x).
-         */
         mounted() {
-            this.prepareComponent();
+            this.getClients();
         },
 
         methods: {
-            /**
-             * Prepare the component.
-             */
-            prepareComponent() {
-                this.getClients();
-
-                $('#modal-create-client').on('shown.bs.modal', () => {
-                    $('#create-client-name').focus();
-                });
-
-                $('#modal-edit-client').on('shown.bs.modal', () => {
-                    $('#edit-client-name').focus();
-                });
-            },
-
-            /**
-             * Get all of the OAuth clients for the user.
-             */
             getClients() {
                 axios.get('/oauth/clients')
                         .then(response => {
@@ -273,48 +221,39 @@
                         });
             },
 
-            /**
-             * Show the form for creating new clients.
-             */
             showCreateClientForm() {
-                $('#modal-create-client').modal('show');
+                this.showCreateModal = true;
+                this.$nextTick(() => {
+                    this.$refs.createNameInput?.focus();
+                });
             },
 
-            /**
-             * Create a new OAuth client for the user.
-             */
             store() {
                 this.persistClient(
                     'post', '/oauth/clients',
-                    this.createForm, '#modal-create-client'
+                    this.createForm, 'showCreateModal'
                 );
             },
 
-            /**
-             * Edit the given client.
-             */
             edit(client) {
                 this.editForm.id = client.id;
                 this.editForm.name = client.name;
                 this.editForm.redirect = client.redirect;
 
-                $('#modal-edit-client').modal('show');
+                this.showEditModal = true;
+                this.$nextTick(() => {
+                    this.$refs.editNameInput?.focus();
+                });
             },
 
-            /**
-             * Update the client being edited.
-             */
             update() {
                 this.persistClient(
                     'put', '/oauth/clients/' + this.editForm.id,
-                    this.editForm, '#modal-edit-client'
+                    this.editForm, 'showEditModal'
                 );
             },
 
-            /**
-             * Persist the client to storage using the given form.
-             */
-            persistClient(method, uri, form, modal) {
+            persistClient(method, uri, form, modalKey) {
                 form.errors = [];
 
                 axios[method](uri, form)
@@ -325,20 +264,17 @@
                         form.redirect = '';
                         form.errors = [];
 
-                        $(modal).modal('hide');
+                        this[modalKey] = false;
                     })
                     .catch(error => {
                         if (typeof error.response.data === 'object') {
-                            form.errors = _.flatten(_.toArray(error.response.data.errors));
+                            form.errors = Object.values(error.response.data.errors).flat();
                         } else {
                             form.errors = ['Something went wrong. Please try again.'];
                         }
                     });
             },
 
-            /**
-             * Destroy the given client.
-             */
             destroy(client) {
                 axios.delete('/oauth/clients/' + client.id)
                         .then(response => {
