@@ -28,7 +28,7 @@
         onload="this.rel = 'stylesheet'"
     >
 
-    <link href="{{ mix('css/app.css') }}" rel="preload stylesheet" as="style">
+    @vite(['resources/assets/sass/app.scss', 'resources/assets/js/app.js'])
 
 
 
@@ -43,14 +43,14 @@
                     {{-- {{ site_title() }} --}}
                     <img src="/images/clingen_logo_75w.png">
                 </a>
-                <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+                <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
                     <span class="navbar-toggler-icon"></span>
                 </button>
 
                 <div class="collapse navbar-collapse" id="navbarSupportedContent">
                     <!-- Left Side Of Navbar -->
                     @if (!Auth::guest())
-                    <ul class="navbar-nav mr-auto">
+                    <ul class="navbar-nav me-auto">
                         @if (Auth::user()->can('create curations'))
                             <li>
                                 <a class="nav-link" href="/#/">Dashboard</a>
@@ -70,7 +70,7 @@
                             <a class="nav-link" href="/#/curations/export">Curation Export</a>
                         </li>
                         <li class="nav-item dropdown">
-                            <a class="nav-link dropdown-toggle" href="#" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                            <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                                 Bulk Lookup <span class="caret"></span>
                             </a>
                             <div class="dropdown-menu" aria-labelledby="nvabarDropdown">
@@ -82,13 +82,13 @@
                     @endif
 
                     <!-- Right Side Of Navbar -->
-                    <ul class="navbar-nav ml-auto">
+                    <ul class="navbar-nav ms-auto">
                         <!-- Authentication Links -->
                         @guest
                             <li><a class="nav-link" href="{{ route('login') }}">Login</a></li>
                         @else
                             <li class="nav-item dropdown">
-                                <a class="nav-link dropdown-toggle" href="#" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                                <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                                     {{ Auth::user()->name }} <span class="caret"></span>
                                 </a>
                                 <div class="dropdown-menu" aria-labelledby="navbarDropdown">
@@ -145,7 +145,7 @@
         window.maxUploadSize = '{{getMaxUploadSizeForHumans()}}'
         window.supportedMimes = {!! json_encode(config('project.supported-mimes')) !!}
     </script>
-    <script src="{{ mix('js/app.js') }}"></script>
+    {{-- JS is loaded via @vite directive in <head> --}}
 
     @stack('scripts')
 </body>
