@@ -1,14 +1,12 @@
 <li class="nav-item" style="margin-right: -46px">
-    <a class="nav-link btn-help" href="#get-help" title="Get help or report a problem" data-toggle="modal" data-target="#help-modal">?</a>
+    <a class="nav-link btn-help" href="#get-help" title="Get help or report a problem" data-bs-toggle="modal" data-bs-target="#help-modal">?</a>
 </li>
 <div class="modal fade" tabindex="-1" role="dialog" id="help-modal">
   <div class="modal-dialog" role="document">
     <div class="modal-content">
       <div class="modal-header">
         <h5 class="modal-title">How can we help?</h5>
-        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-          <span aria-hidden="true">&times;</span>
-        </button>
+        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
       </div>
       <div class="modal-body">
         <p class="lead text-center">
